+++ tests/Pgsql/SelectTest.php
@@ -6,4 +6,117 @@ use Aura\SqlQuery\Common;
 class SelectTest extends Common\SelectTest
 {
     protected $db_type = 'pgsql';
+
+    public function testLateralJoinSubSelect()
+    {
+        $this->query->cols(['*']);
+        $this->query->from('t1');
+        $this->query->lateralJoinSubSelect(
+            'left',
+            'SELECT * FROM t2 WHERE t2.c1 = t1.c1',
+            'a2',
+            't2.c1 = a2.c1'
+        );
+        $expect = '
+            SELECT
+                *
+            FROM
+                <<t1>>
+                    LEFT JOIN LATERAL (
+                        SELECT * FROM t2 WHERE t2.c1 = t1.c1
+                    ) <<a2>> ON <<t2>>.<<c1>> = <<a2>>.<<c1>>
+        ';
+        $actual = $this->query->__toString();
+        $this->assertSameSql($expect, $actual);
+    }
+
+    public function testLateralJoinSubSelect_selectObject()
+    {
+        $sub = $this->newQuery();
+        $sub->cols(['*'])->from('t2')->where('t2.c2 = :v', ['v' => 9]);
+
+        $this->query->cols(['*']);
+        $this->query->from('t1');
+        $this->query->lateralJoinSubSelect('inner', $sub, 'a2', 't2.c1 = a2.c1');
+
+        $expect = '
+            SELECT
+                *
+            FROM
+                <<t1>>
+                    INNER JOIN LATERAL (
+                        SELECT
+                            *
+                        FROM
+                            <<t2>>
+                        WHERE
+                            <<t2>>.<<c2>> = :v
+                    ) <<a2>> ON <<t2>>.<<c1>> = <<a2>>.<<c1>>
+        ';
+        $actual = $this->query->__toString();
+        $this->assertSameSql($expect, $actual);
+
+        $this->assertSame(['v' => 9], $this->query->getBindValues());
+    }
+
+    /**
+     * Postgres requires an ON clause on LEFT/INNER JOIN LATERAL; without a
+     * condition the query is a syntax error, so default to ON true.
+     */
+    public function testLateralJoinSubSelect_noCondition()
+    {
+        $this->query->cols(['*']);
+        $this->query->from('t1');
+        $this->query->lateralJoinSubSelect(
+            'left',
+            'SELECT * FROM t2 WHERE t2.c1 = t1.c1',
+            'a2'
+        );
+        $expect = '
+            SELECT
+                *
+            FROM
+                <<t1>>
+                    LEFT JOIN LATERAL (
+                        SELECT * FROM t2 WHERE t2.c1 = t1.c1
+                    ) <<a2>> ON true
+        ';
+        $actual = $this->query->__toString();
+        $this->assertSameSql($expect, $actual);
+    }
+
+    /**
+     * CROSS JOIN LATERAL takes no ON clause at all.
+     */
+    public function testLateralJoinSubSelect_cross()
+    {
+        $this->query->cols(['*']);
+        $this->query->from('t1');
+        $this->query->lateralJoinSubSelect(
+            'cross',
+            'SELECT * FROM t2 WHERE t2.c1 = t1.c1',
+            'a2'
+        );
+        $expect = '
+            SELECT
+                *
+            FROM
+                <<t1>>
+                    CROSS JOIN LATERAL (
+                        SELECT * FROM t2 WHERE t2.c1 = t1.c1
+                    ) <<a2>>
+        ';
+        $actual = $this->query->__toString();
+        $this->assertSameSql($expect, $actual);
+    }
+
+    public function testLateralJoinSubSelect_duplicateRef()
+    {
+        $this->query->cols(['*']);
+        $this->query->from('t1');
+        $this->query->lateralJoinSubSelect('left', 'SELECT * FROM t2', 'a2', 't2.c1 = a2.c1');
+
+        $this->expectException(\Aura\SqlQuery\Exception\LogicException::class);
+        $this->query->lateralJoinSubSelect('left', 'SELECT * FROM t3', 'a2', 't3.c1 = a2.c1');
+    }
 }